improve CCache types
- Use size_t to represent sizes.
  As most of C size function return this type.
- Use explicit casting when types change.
- Avoid float type for expression result integers.
- remove `int` from `long`, `short` and `unsigned`.

+++ CCache/snprintf.c
@@ -259,7 +259,7 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 				if (cflags == DP_C_SHORT) 
 					value = va_arg (args, int);
 				else if (cflags == DP_C_LONG)
-					value = va_arg (args, long int);
+					value = va_arg (args, long);
 				else if (cflags == DP_C_LLONG)
 					value = va_arg (args, LLONG);
 				else
@@ -269,25 +269,25 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 			case 'o':
 				flags |= DP_F_UNSIGNED;
 				if (cflags == DP_C_SHORT)
-					value = va_arg (args, unsigned int);
+					value = va_arg (args, unsigned);
 				else if (cflags == DP_C_LONG)
-					value = (long)va_arg (args, unsigned long int);
+					value = (long)va_arg (args, unsigned long);
 				else if (cflags == DP_C_LLONG)
 					value = (long)va_arg (args, unsigned LLONG);
 				else
-					value = (long)va_arg (args, unsigned int);
+					value = (long)va_arg (args, unsigned);
 				fmtint (buffer, &currlen, maxlen, value, 8, min, max, flags);
 				break;
 			case 'u':
 				flags |= DP_F_UNSIGNED;
 				if (cflags == DP_C_SHORT)
-					value = va_arg (args, unsigned int);
+					value = va_arg (args, unsigned);
 				else if (cflags == DP_C_LONG)
-					value = (long)va_arg (args, unsigned long int);
+					value = (long)va_arg (args, unsigned long);
 				else if (cflags == DP_C_LLONG)
 					value = (LLONG)va_arg (args, unsigned LLONG);
 				else
-					value = (long)va_arg (args, unsigned int);
+					value = (long)va_arg (args, unsigned);
 				fmtint (buffer, &currlen, maxlen, value, 10, min, max, flags);
 				break;
 			case 'X':
@@ -296,13 +296,13 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 			case 'x':
 				flags |= DP_F_UNSIGNED;
 				if (cflags == DP_C_SHORT)
-					value = va_arg (args, unsigned int);
+					value = va_arg (args, unsigned);
 				else if (cflags == DP_C_LONG)
-					value = (long)va_arg (args, unsigned long int);
+					value = (long)va_arg (args, unsigned long);
 				else if (cflags == DP_C_LLONG)
 					value = (LLONG)va_arg (args, unsigned LLONG);
 				else
-					value = (long)va_arg (args, unsigned int);
+					value = (long)va_arg (args, unsigned);
 				fmtint (buffer, &currlen, maxlen, value, 16, min, max, flags);
 				break;
 			case 'f':
@@ -338,7 +338,7 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 				strvalue = va_arg (args, char *);
 				if (!strvalue) strvalue = "(NULL)";
 				if (max == -1) {
-					max = strlen(strvalue);
+					max = (int)strlen(strvalue);
 				}
 				if (min > 0 && max >= 0 && min > max) max = min;
 				fmtstr (buffer, &currlen, maxlen, strvalue, flags, min, max);
@@ -349,13 +349,13 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 				break;
 			case 'n':
 				if (cflags == DP_C_SHORT) {
-					short int *num;
+					short *num;
 					num = va_arg (args, short int *);
-					*num = currlen;
+					*num = (short)currlen;
 				} else if (cflags == DP_C_LONG) {
-					long int *num;
-					num = va_arg (args, long int *);
-					*num = (long int)currlen;
+					long *num;
+					num = va_arg (args, long *);
+					*num = (long)currlen;
 				} else if (cflags == DP_C_LLONG) {
 					LLONG *num;
 					num = va_arg (args, LLONG *);
@@ -363,7 +363,7 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 				} else {
 					int *num;
 					num = va_arg (args, int *);
-					*num = currlen;
+					*num = (int)currlen;
 				}
 				break;
 			case '%':
@@ -756,7 +756,7 @@ static void dopr_outch(char *buffer, size_t *currlen, size_t maxlen, char c)
 #if !defined(HAVE_VSNPRINTF) || !defined(HAVE_C99_VSNPRINTF)
  int vsnprintf (char *str, size_t count, const char *fmt, va_list args)
 {
-	return dopr(str, count, fmt, args);
+	return (int)dopr(str, count, fmt, args);
 }
 #endif
